abstracted and re-used option processing logic
