+++ src/cli/lib/handle-create.mjs
@@ -20,6 +20,7 @@ const handleCreate = async ({ argv, globalOptions, sitesInfo }) => {
   const sourcePath = fsPath.resolve(createOptions['source-path'])
   let sourceType = createOptions['source-type']
   const stackName = createOptions['stack-name']
+  const options = optionsLib.mapRawOptions(setOptionOptions.options)
 
   const siteInfo = sitesInfo[apexDomain] || { apexDomain, bucketName, sourcePath, sourceType }
   siteInfo.region = createOptions.region || siteInfo.region || 'us-east-1'
@@ -51,6 +52,8 @@ const handleCreate = async ({ argv, globalOptions, sitesInfo }) => {
     process.stderr(`Invalid bucket name. Must be valid AWS S3 Transfer Accelerated bucket name matching: ${awsS3TABucketNameREString}`)
     process.exit(2) // eslint-disable-line no-process-exit
   }
+
+  optionLib.updatePluginSettings({ options, siteInfo })
 
   // update siteInfo in case these were manually specified
   for (const [value, field] of [[bucketName, 'bucketName'], [sourcePath, 'sourcePath'], [sourceType, 'sourceType']]) {
